+++ packages/ui-radio/package.json
@@ -2,7 +2,8 @@
   "name": "@halvaradop/ui-radio",
   "version": "0.2.0",
   "private": false,
-  "description": "A customizable Radio component for @halvaradop/ui library with Tailwind CSS styling.",
+  "description": "DEPRECATED: Use @halvaradop/ui-radio-group instead. This package will no longer receive updates.",
+  "deprecated": true,
   "type": "module",
   "types": "dist/index.d.ts",
   "scripts": {
