+++ packages/ui-radio/src/index.ts
@@ -1,0 +1,13 @@
+/**
+ * @package @halvaradop/ui-radio
+ *
+ * @deprecated This package is deprecated and will no longer receive updates or bug fixes.
+ * Please migrate to @halvaradop/ui-radio-group.
+ *
+ * For more information, see:
+ * https://github.com/halvaradop/ui/issues/144
+ */
+
+throw new Error(
+    "[DEPRECATED] @halvaradop/ui-radio: This package is deprecated. Please use @halvaradop/ui-radio-group instead. See https://github.com/halvaradop/ui/issues/144"
+)
